api/requirements.txt
azure-cosmos==4.5.0
azure-identity==1.12.0
requests==2.28.2
opencensus-ext-azure==1.1.9
opencensus-ext-logging==0.1.1
opencensus-ext-flask==0.8.1
adal==1.2.7
azure-keyvault==4.2.0
fastapi==0.101.0
uvicorn==0.23.2
fastapi-azure-auth==4.2.1
